+++ tests/conftest.py
@@ -1,0 +1,8 @@
+import logging
+
+import pytest
+
+
+@pytest.fixture
+def debugger(caplog):
+    caplog.set_level(logging.DEBUG)